Introduce a ref_time property for VDIF headers.
This in analogy with mark4 and mark 5b, and exposing the 'ref_epoch' key.
The main reason is that generally when setting the time in a (copy of a)
header, one does *not* want to change the reference epoch, but this means
it should be possible to set it more easily in fromvalues.