Assume properties are on header0 by default.
This reduces the need to pass on properties. One could even try to define
them all as being on the header by default, and override definitions in
other classes, but beyond not being very backwards compatible, it also
turned out to be rather unhandy for sample_rate, which sometimes is best
overridden.  More generally it assumes all properties of a header are
OK, which may not be the case.